evix-protocol: move wire types out of evix engine
