+++ crates/evix-protocol/Cargo.toml
@@ -8,6 +8,5 @@ license.workspace      = true
 repository.workspace   = true
 
 [dependencies]
-evix.workspace       = true
 serde.workspace      = true
 serde_json.workspace = true
