+++ crates/evix-protocol/src/serde_config.rs
@@ -115,63 +115,3 @@ pub mod auto_args {
     )
   }
 }
-
-#[cfg(test)]
-mod tests {
-  use std::path::PathBuf;
-
-  use crate::{AutoArg, Config, Input};
-
-  #[test]
-  fn config_uses_daemon_wire_shape() {
-    let config = Config {
-      input: Input::Flake(".#jobs".into()),
-      auto_args: vec![
-        ("pkgs".into(), AutoArg::Expr("import <nixpkgs> {}".into())),
-        ("name".into(), AutoArg::Str("hello".into())),
-      ],
-      max_memory_size: 128,
-      ..Config::default()
-    };
-
-    let json = serde_json::to_value(&config).unwrap();
-    assert_eq!(
-      json["input"],
-      serde_json::json!({"type": "flake", "value": ".#jobs"})
-    );
-    assert_eq!(json["maxMemorySize"], 128);
-    assert_eq!(
-      json["autoArgs"],
-      serde_json::json!([
-        {"name": "pkgs", "kind": "expr", "value": "import <nixpkgs> {}"},
-        {"name": "name", "kind": "str", "value": "hello"}
-      ])
-    );
-
-    let roundtrip: Config = serde_json::from_value(json).unwrap();
-    let Input::Flake(input) = roundtrip.input else {
-      panic!("expected flake input");
-    };
-    assert_eq!(input, ".#jobs");
-    assert_eq!(roundtrip.auto_args.len(), 2);
-  }
-
-  #[test]
-  fn file_input_roundtrips() {
-    let config = Config {
-      input: Input::File(PathBuf::from("default.nix")),
-      ..Config::default()
-    };
-
-    let json = serde_json::to_value(&config).unwrap();
-    assert_eq!(
-      json["input"],
-      serde_json::json!({"type": "file", "path": "default.nix"})
-    );
-    let roundtrip: Config = serde_json::from_value(json).unwrap();
-    let Input::File(path) = roundtrip.input else {
-      panic!("expected file input");
-    };
-    assert_eq!(path, PathBuf::from("default.nix"));
-  }
-}
